requirements.txt
# VillageCrop - Unified Application Requirements

# Web Framework
Flask==3.0.0
Flask-PyMongo==3.0.1
Flask-CORS==4.0.0

# Database
pymongo==4.6.0
bcrypt==4.1.2

# Image Processing
Pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.26.2

# ML/DL
torch==2.1.1
torchvision==0.16.1

# Utilities
python-dotenv==1.0.0
geojson==3.1.0
requests==2.31.0
Werkzeug==3.0.1
